+++ login.php
@@ -1,709 +1,719 @@
 ﻿<!DOCTYPE html>
-	<html>
-		<head>
-			<meta charset="utf-8">
-			<title>Вход</title>
-			<meta name="description" content="Центральный украинский ресурс по купле / продаже неликвидов">
-			<meta name="keywords" content="Неликвиды, купля, продажа">
-			<meta name="author" content="SITE&SEO">
-			<link rel="stylesheet" href="css/normalize.css">
-			<link rel="stylesheet" href="css/bootstrap.css">
-			<link rel="stylesheet" href="css/font-awesome-4.2.0/css/font-awesome.min.css">
-			<link rel="stylesheet" href="css/perfect-scrollbar.css">
-			<link rel="stylesheet" href="css/jquery.fancybox.css">
-			<link href='http://fonts.googleapis.com/css?family=Open+Sans' rel='stylesheet' type='text/css'>
-			<link href='http://fonts.googleapis.com/css?family=Roboto' rel='stylesheet' type='text/css'>
-			<link rel="stylesheet" href="css/style.css">
-			<link rel="stylesheet" href="css/responsive.css">
-		</head>
-		<body class="no-bg"> 
-		
-			
-			<section id="page">
-			
-		
-				
-				<header id="header" class="normal-header line-decoration">
-				
-					<!-- Top Bar -->
-					<div id="top-bar" class="text-normal full-width-line clearfix">
-						<div class="wf-wrap">
-							<div class="wf-container-top">
-								<div class="wf-table">
-									<!-- Top bar conctacts icons -->
-									<div class="wf-td">
-										<span class="mini-contacts email">
-											<i class="fa fa-envelope-o"></i>
-											<span class="top-bar-icon-text">isitinme88@gmail.com</span>
-										</span>
-										<span class="mini-contacts phone">
-											<i class="fa fa-phone"></i>
-											<span class="top-bar-icon-text">(095) 896 24 70</span>
-										</span>
-										<span class="mini-contacts skype">
-											<i class="fa fa-skype"></i>
-											<span class="top-bar-icon-text">alexander_punkk</span>
-										</span>
-									</div>
-									<!-- Top bar conctacts icons End -->
-									
-									<div class="right-block wf-td clearfix">
-										<div class="mini-login">
-											<a href="login.php" class="submit">
-												<i class="fa fa-sign-in"></i>
-												<span class="top-bar-icon-text">Войти</span>
-											</a>
-										</div>
-									
-										<div class="mini-login">
-											<a href="registration.php" class="submit">
-												<i class="fa fa-pencil"></i>
-												<span class="top-bar-icon-text">Зарегистрироваться</span>
-											</a>
-										</div>
-										<!-- Social -->
-										<div class="soc-ico clearfix">
-											<a href="#" title="Instagram" target="_blank">
-												<i class="fa fa-instagram"></i>
-											</a>
-											<a href="#" title="Facebook" target="_blank">
-												<i class="fa fa-facebook-square"></i>
-											</a>
-											<a href="#" title="Twitter" target="_blank">
-												<i class="fa fa-twitter"></i>
-											</a>
-											<a href="#" title="Vk" target="_blank">
-												<i class="fa fa-vk"></i>
-											</a>
-										</div>
-										<!-- Social End -->
-									</div>
-									
-								</div>
-							</div>
-						</div>
-					</div>
-					
-					<span class="top-bar-arrow">
-						<span>
-							<i class="fa fa-angle-down"></i>
-						</span>
-					</span>
-					
-					<!-- Top Bar End -->
-					
-					<!-- Branding & Navigation -->
-					<div class="wf-wrap clearfix">
-							
-						<div id="branding" class="wf-td">
-							<a href="index.php">
-								<img src="img/logo-regular.png" alt="Logo">
-							</a>
-						</div>
-						
-						
-						<nav id="navigation" class="wf-td">
-							<!-- Mobile nav -->
-							<div id="dl-menu">
-								<a href="#" id="mobile-menu">
-									<div class="mobile-button">
-										<span class="menu-open" style="display: inline;">Menu</span>
-										<span class="menu-back">Back</span>
-									</div>
-								</a>
-								<div class="dl-container">
-									<ul class="dl-menu">
-										<li class="act">
-											<a href="#" onclick="return false;">
-												<span class="dl-text">Home</span>
-												<span class="dl-icon">
-													<i class="fa fa-arrow-right"></i>
-												</span>
-											</a>
-										</li>
-										<li>
-											<a href="#" onclick="return false;">
-												<span class="dl-text">Pages</span>
-												<span class="dl-icon">
-													<i class="fa fa-arrow-right"></i>
-												</span>
-											</a>
-										</li>
-										<li>
-											<a href="#" onclick="return false;">
-												<span class="dl-text">Blog</span>
-												<span class="dl-icon">
-													<i class="fa fa-arrow-right"></i>
-												</span>
-											</a>
-										</li>
-										<li>
-											<a href="#" onclick="return false;">
-												<span class="dl-text">Portfolio</span>
-												<span class="dl-icon">
-													<i class="fa fa-arrow-right"></i>
-												</span>
-											</a>
-										</li>
-										<li>
-											<a href="#" onclick="return false;">
-												<span class="dl-text">Gallery</span>
-												<span class="dl-icon">
-													<i class="fa fa-arrow-right"></i>
-												</span>
-											</a>
-										</li>
-									</ul>
-								</div>
-							</div>
-						
-							<!-- Main Nav -->
-							<ul id="main-nav" class="clearfix">
-								<li class="act no-link">
-									<a href="#" onclick="return false;" class="topItem">
-										<span>
-											Home
-										</span>
-									</a>
-									<div class="sub-nav">
-										<ul>
-											<li>
-												<a href="index.php" onclick="return false;">
-													<span>1. Business</span>
-												</a>
-											</li>
-											<li>
-												<a href="index.php" onclick="return false;">
-													<span>2. Presentation</span>
-												</a>
-											</li>
-											<li>
-												<a href="index.php" onclick="return false;">
-													<span>3. Creative</span>
-												</a>
-											</li>
-											<li>
-												<a href="index.php" onclick="return false;">
-													<span>4. Photography</span>
-												</a>
-											</li>
-										</ul>
-									</div>
-								</li>
-								<li class="no-link">
-									<a href="#" onclick="return false;" class="topItem">
-										<span>
-											Pages
-											<i class="underline"></i>
-										</span>
-									</a>
-									<div class="sub-nav">
-										<ul>
-											<li>
-												<a href="#" onclick="return false;">
-													<span>1. Business</span>
-												</a>
-											</li>
-											<li>
-												<a href="#" onclick="return false;">
-													<span>2. Presentation</span>
-												</a>
-											</li>
-											<li>
-												<a href="#" onclick="return false;">
-													<span>3. Creative</span>
-												</a>
-											</li>
-											<li>
-												<a href="#" onclick="return false;">
-													<span>4. Photography</span>
-												</a>
-											</li>
-										</ul>
-									</div>
-								</li>
-								<li class="no-link">
-									<a href="#" onclick="return false;" class="topItem">
-										<span>
-											Blog
-											<i class="underline"></i>
-										</span>
-									</a>
-									<div class="sub-nav">
-										<ul>
-											<li>
-												<a href="#" onclick="return false;">
-													<span>1. Business</span>
-												</a>
-											</li>
-											<li>
-												<a href="#" onclick="return false;">
-													<span>2. Presentation</span>
-												</a>
-											</li>
-											<li>
-												<a href="#" onclick="return false;">
-													<span>3. Creative</span>
-												</a>
-											</li>
-											<li>
-												<a href="#" onclick="return false;">
-													<span>4. Photography</span>
-												</a>
-											</li>
-										</ul>
-									</div>
-								</li>
-								<li class="no-link">
-									<a href="#" onclick="return false;" class="topItem">
-										<span>
-											Portfolio
-											<i class="underline"></i>
-										</span>
-									</a>
-									<div class="sub-nav">
-										<ul>
-											<li>
-												<a href="#" onclick="return false;">
-													<span>1. Business</span>
-												</a>
-											</li>
-											<li>
-												<a href="#" onclick="return false;">
-													<span>2. Presentation</span>
-												</a>
-											</li>
-											<li>
-												<a href="#" onclick="return false;">
-													<span>3. Creative</span>
-												</a>
-											</li>
-											<li>
-												<a href="#" onclick="return false;">
-													<span>4. Photography</span>
-												</a>
-											</li>
-										</ul>
-									</div>
-								</li>
-								<li class="no-link">
-									<a href="#" onclick="return false;" class="topItem">
-										<span>
-											Gallery
-											<i class="underline"></i>
-										</span>
-									</a>
-									<div class="sub-nav">
-										<ul>
-											<li>
-												<a href="#" onclick="return false;">
-													<span>1. Business</span>
-												</a>
-											</li>
-											<li>
-												<a href="#" onclick="return false;">
-													<span>2. Presentation</span>
-												</a>
-											</li>
-											<li>
-												<a href="#" onclick="return false;">
-													<span>3. Creative</span>
-												</a>
-											</li>
-											<li>
-												<a href="#" onclick="return false;">
-													<span>4. Photography</span>
-												</a>
-											</li>
-										</ul>
-									</div>
-								</li>
-							</ul>
-						</nav>
-							
-					</div>
-					<!-- Branding & Navigation End -->
-					
-				</header>
-				
-				<!-- Header End -->
-				
-				<!-- Page title -->
-				
-				<div class="page-title">
-					<div class="wf-wrap">
-						
-						<div class="wf-table">
-							<div class="wf-td hgroup">
-								<h1>Вход</h1>
-							</div>
-							<div class="wf-td">
-								<ul class="breadcrumbs text-normal">
-									<li>
-										<a href="index.php">Главная</a>
-									</li>
-									<li>
-										<a href="login.php">Вход</a>
-									</li>
-								</ul>
-							</div>
-						</div>
-						
-					</div>
-				</div>
-				
-				<!-- Page title -->
-				
-				<!-- Main Content -->
-				
-				<div id="main">
-					<div class="container wf-wrap">
-						<div class="row">
-							
-							<div class="form-container">
-								<form name="login-form" action="#" method="#">
-									<div class="form-fields">
-										<span class="form-name login-form-input">
-											<input type="text" class="validate" placeholder="E-mail *" name="name" value="">
-										</span>
-										<span class="form-name login-form-input">
-											<input type="text" class="validate" placeholder="Пароль *" name="name" value="">
-										</span>
-									</div>
-									
-									<p class="remember-me">
-										<label>
-											<input name="rememberme" type="checkbox" id="rememberme" value="">
-											Запомнить меня
-										</label>
-									</p>
-									
-									<p>
-										<a href="cabinet.php" id="doRegistr" class="form-submit">
-											<span>Войти</span>
-										</a>
-									</p>
-								</form>
-							</div>
-							
-						</div>
-						
-						<hr>
-						
-						<div class="row">
-							<!-- Feedback Section -->
-							
-							<div id="feedback-container">
-							
-								<header>
-									<h2>Наши клиенты и партнеры</h2>
-									<h4>Нас уже более 10 000!</h4>
-								</header>
-							
-								<div class="feeback-content clearfix">
-								
-									<div id="feedback-carousel-comments" class="col-sm-6">
-									
-										<!-- Customer comments block -->
-										<section id="testimonial-slider" class="carousel slide">
-										
-												<ol class="carousel-indicators">
-													<li data-target="#testimonial-slider" data-slide-to="0" class="active"></li>
-													<li data-target="#testimonial-slider" data-slide-to="1"></li>
-													<li data-target="#testimonial-slider" data-slide-to="2"></li>
-												</ol>
-												
-												<div class="carousel-inner" role="listbox">
-													<div class="item active">
-														<p>
-															Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris ipsum erat, finibus sit amet fringilla id, accumsan sed nisl. 
-															Nulla odio eros, blandit ac metus faucibus, sollicitudin posuere sapien. Maecenas ut convallis arcu. Phasellus at tellus 
-															sed odio vestibulum sodales in at lacus. Sed commodo metus et sapien pretium, ac sollicitudin enim dignissim. Donec tempus 
-															diam et porta aliquam. Ut efficitur sollicitudin diam a accumsan.
-														</p>
-														<span class="img">
-															<img src="img/feedback-person-1.jpg" alt="person 1">
-														</span>
-														<span class="who">
-															<span class="text-primary">Diana Richards</span>
-															<br>
-															<span class="text-secondary">manager</span>
-														</span>
-													</div>
-													<div class="item">
-														<p>
-															Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris ipsum erat, finibus sit amet fringilla id, accumsan sed nisl. 
-															Nulla odio eros, blandit ac metus faucibus, sollicitudin posuere sapien. Maecenas ut convallis arcu. Phasellus at tellus 
-															sed odio vestibulum sodales in at lacus. Sed commodo metus et sapien pretium, ac sollicitudin enim dignissim. Donec tempus 
-															diam et porta aliquam. Ut efficitur sollicitudin diam a accumsan.
-														</p>
-														<span class="img">
-															<img src="img/feedback-person-2.jpg" alt="person 2">
-														</span>
-														<span class="who">
-															<span class="text-primary">Anna White</span>
-															<br>
-															<span class="text-secondary">loyal client</span>
-														</span>
-													</div>
-													<div class="item">
-														<p>
-															Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris ipsum erat, finibus sit amet fringilla id, accumsan sed nisl. 
-															Nulla odio eros, blandit ac metus faucibus, sollicitudin posuere sapien. Maecenas ut convallis arcu. Phasellus at tellus 
-															sed odio vestibulum sodales in at lacus. Sed commodo metus et sapien pretium, ac sollicitudin enim dignissim. Donec tempus 
-															diam et porta aliquam. Ut efficitur sollicitudin diam a accumsan.
-														</p>
-														<span class="img">
-															<img src="img/feedback-person-3.jpg" alt="person 3">
-														</span>
-														<span class="who">
-															<span class="text-primary">Jacob Firebird</span>
-															<br>
-															<span class="text-secondary">happy client</span>
-														</span>
-													</div>
-												</div>
-												
-										</section>
-										<!-- Customer comments block End -->
-									</div>
-									
-									<div id="feedback-partners" class="col-sm-6">
-										<!-- Partners icons block -->
-										<section id="partners-icons" class="clearfix">
-										
-											<div class="col-sm-4 partner-icon">
-												<a href="#" target="_blank" title="Partner 1">
-													<img src="img/l-authentic.png">
-												</a>
-											</div>
-											
-											<div class="col-sm-4 partner-icon">
-												<a href="#" target="_blank" title="Partner 2">
-													<img src="img/l-hamb.png">
-												</a>
-											</div>
-											
-											<div class="col-sm-4 partner-icon">
-												<a href="#" target="_blank" title="Partner 3">
-													<img src="img/l-ransom.png">
-												</a>
-											</div>
-											
-											<div class="col-sm-4 partner-icon">
-												<a href="#" target="_blank" title="Partner 3">
-													<img src="img/l-steak.png">
-												</a>
-											</div>
-											
-											<div class="col-sm-4 partner-icon">
-												<a href="#" target="_blank" title="Partner 3">
-													<img src="img/l-thesaviour.png">
-												</a>
-											</div>
-											
-											<div class="col-sm-4 partner-icon">
-												<a href="#" target="_blank" title="Partner 3">
-													<img src="img/l-vintage.png">
-												</a>
-											</div>
-										
-										</section>
-										<!-- Partners icons block End -->
-									</div>
-								
-								</div>
-								
-							</div>
-					
-							<!-- Feedback Section End -->
-						</div>
-						
-						
-					</div>
-				</div>
-				
-				<!-- Main Content End -->
-				
-				<footer id="footer">
-				
-					<div class="wf-wrap">
-						<div class="wf-container-footer">
-							<div class="wf-container clearfix">
-								
-								<section class="widget col-sm-4" id="footer-contact-us-widget">
-									<div class="widget-title">НАШИ КОНТАКТЫ</div>
-									<ul class="contact-info">
-										<li>
-											<span class="color-primary">Адрес:</span>
-											<br>
-											8500 Beverly Boulevard
-											<br>
-											Los Angeles, CA 90048
-										</li>
-										<li>
-											<span class="color-primary">Телефоны:</span>
-											<br>
-											+12 345 67 00 89
-											<br>
-											+12 987 00 65 43
-										</li>
-										<li>
-											<span class="color-primary">E-mail:</span>
-											<br>
-											isitinme88@gmail.com
-										</li>
-									</ul>
-									<div class="soc-ico clearfix">
-										<a href="#" title="Instagram" target="_blank">
-											<i class="fa fa-instagram"></i>
-										</a>
-										<a href="#" title="Facebook" target="_blank">
-											<i class="fa fa-facebook-square"></i>
-										</a>
-										<a href="#" title="Twitter" target="_blank">
-											<i class="fa fa-twitter"></i>
-										</a>
-										<a href="#" title="Vk" target="_blank">
-											<i class="fa fa-vk"></i>
-										</a>
-									</div>
-								</section>
-								
-								<section class="widget col-sm-4" id="footer-our-services-widget">
-									<div class="widget-title">НАШИ ПРЕДЛОЖЕНИЯ</div>
-								
-									<ul class="foot-accordion clearfix">
-										<li>
-											<a href="#" class="text-primary">
-												<span class="accor-toggle-icon">
-													<i class="fa fa-angle-down"></i>
-												</span>
-												<span class="color-primary">Web Design</span>
-											</a>
-											<div class="accor-content">
-												Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris ipsum erat, finibus sit amet fringilla id, accumsan sed nisl. 
-											</div>
-										</li>
-										<li>
-											<a href="#" class="text-primary">
-												<span class="accor-toggle-icon">
-													<i class="fa fa-angle-down"></i>
-												</span>
-												<span class="color-primary">Photography</span>
-											</a>
-											<div class="accor-content">
-												Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris ipsum erat, finibus sit amet fringilla id, accumsan sed nisl. 
-											</div>
-										</li>
-										<li>
-											<a href="#" class="text-primary">
-												<span class="accor-toggle-icon">
-													<i class="fa fa-angle-down"></i>
-												</span>
-												<span class="color-primary">Programming</span>
-											</a>
-											<div class="accor-content">
-												Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris ipsum erat, finibus sit amet fringilla id, accumsan sed nisl. 
-											</div>
-										</li>
-										<li>
-											<a href="#" class="text-primary">
-												<span class="accor-toggle-icon">
-													<i class="fa fa-angle-down"></i>
-												</span>
-												<span class="color-primary">Marketing & PR</span>
-											</a>
-											<div class="accor-content">
-												Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris ipsum erat, finibus sit amet fringilla id, accumsan sed nisl. 
-											</div>
-										</li>
-									</ul>
-								
-								</section>
-								
-								<section class="widget col-sm-4" id="footer-featured-posts-widget">
-									<div class="widget-title">ПОПУЛЯРНЫЕ СТАТЬИ</div>
-									
-									<ul class="recent-posts">
-										<li>
-											<article>
-												<div class="wf-td">
-													<a href="#" class="alignLeft">
-														<img src="img/article_img_1.jpg" alt="Image" width="40" height="40">
-													</a>
-												</div>
-												<div class="post-content">
-													<a href="#">Business cards design trends</a>
-													<br>
-													<time class="text-secondary">August 5, 2013</time>
-												</div>
-											</article>
-										</li>
-										<li>
-											<article>
-												<div class="wf-td">
-													<a href="#" class="alignLeft">
-														<img src="img/article_img_1.jpg" alt="Image" width="40" height="40">
-													</a>
-												</div>
-												<div class="post-content">
-													<a href="#">Business cards design trends</a>
-													<br>
-													<time class="text-secondary">August 5, 2013</time>
-												</div>
-											</article>
-										</li>
-										<li>
-											<article>
-												<div class="wf-td">
-													<a href="#" class="alignLeft">
-														<img src="img/article_img_1.jpg" alt="Image" width="40" height="40">
-													</a>
-												</div>
-												<div class="post-content">
-													<a href="#">Business cards design trends</a>
-													<br>
-													<time class="text-secondary">August 5, 2013</time>
-												</div>
-											</article>
-										</li>
-										<li>
-											<article>
-												<div class="wf-td">
-													<a href="#" class="alignLeft">
-														<img src="img/article_img_1.jpg" alt="Image" width="40" height="40">
-													</a>
-												</div>
-												<div class="post-content">
-													<a href="#">Business cards design trends</a>
-													<br>
-													<time class="text-secondary">August 5, 2013</time>
-												</div>
-											</article>
-										</li>
-									</ul>
-								</section>
-								
-							</div>
-						</div>
-					</div>
-				
-				</footer>
-				
-				<a href="#" class="scroll-top">
-					<i class="fa fa-arrow-up"></i>
-				</a>
-			
-			</section>
-			
-			
-			<script src="js/jquery-1.11.0.min.js"></script>
-			<script src="js/jquery-ui.js"></script>
-			<script src="js/bootstrap.js"></script>
-			<script src="js/perfect-scrollbar.jquery.js"></script>
-			<script src="js/jquery.fancybox.pack.js"></script>
-			<script src="js/fixedMenu.js"></script>
-			<script src="js/main.js"></script>
-		</body>
-	</html>
+<html>
+    <head>
+        <meta charset="utf-8">
+        <title>Вход</title>
+        <meta name="description" content="Центральный украинский ресурс по купле / продаже неликвидов">
+        <meta name="keywords" content="Неликвиды, купля, продажа">
+        <meta name="author" content="SITE&SEO">
+        <link rel="stylesheet" href="css/normalize.css">
+        <link rel="stylesheet" href="css/bootstrap.css">
+        <link rel="stylesheet" href="css/font-awesome-4.2.0/css/font-awesome.min.css">
+        <link rel="stylesheet" href="css/perfect-scrollbar.css">
+        <link rel="stylesheet" href="css/jquery.fancybox.css">
+        <link href='http://fonts.googleapis.com/css?family=Open+Sans' rel='stylesheet' type='text/css'>
+        <link href='http://fonts.googleapis.com/css?family=Roboto' rel='stylesheet' type='text/css'>
+        <link rel="stylesheet" href="css/style.css">
+        <link rel="stylesheet" href="css/responsive.css">
+    </head>
+    <body class="no-bg"> 
+
+
+        <section id="page">
+
+
+
+            <header id="header" class="normal-header line-decoration">
+
+                <!-- Top Bar -->
+                <div id="top-bar" class="text-normal full-width-line clearfix">
+                    <div class="wf-wrap">
+                        <div class="wf-container-top">
+                            <div class="wf-table">
+                                <!-- Top bar conctacts icons -->
+                                <div class="wf-td">
+                                    <span class="mini-contacts email">
+                                        <i class="fa fa-envelope-o"></i>
+                                        <span class="top-bar-icon-text">isitinme88@gmail.com</span>
+                                    </span>
+                                    <span class="mini-contacts phone">
+                                        <i class="fa fa-phone"></i>
+                                        <span class="top-bar-icon-text">(095) 896 24 70</span>
+                                    </span>
+                                    <span class="mini-contacts skype">
+                                        <i class="fa fa-skype"></i>
+                                        <span class="top-bar-icon-text">alexander_punkk</span>
+                                    </span>
+                                </div>
+                                <!-- Top bar conctacts icons End -->
+
+                                <div class="right-block wf-td clearfix">
+                                    <div class="mini-login">
+                                        <a href="login.php" class="submit">
+                                            <i class="fa fa-sign-in"></i>
+                                            <span class="top-bar-icon-text">Войти</span>
+                                        </a>
+                                    </div>
+
+                                    <div class="mini-login">
+                                        <a href="registration.php" class="submit">
+                                            <i class="fa fa-pencil"></i>
+                                            <span class="top-bar-icon-text">Зарегистрироваться</span>
+                                        </a>
+                                    </div>
+                                    <!-- Social -->
+                                    <div class="soc-ico clearfix">
+                                        <a href="#" title="Instagram" target="_blank">
+                                            <i class="fa fa-instagram"></i>
+                                        </a>
+                                        <a href="#" title="Facebook" target="_blank">
+                                            <i class="fa fa-facebook-square"></i>
+                                        </a>
+                                        <a href="#" title="Twitter" target="_blank">
+                                            <i class="fa fa-twitter"></i>
+                                        </a>
+                                        <a href="#" title="Vk" target="_blank">
+                                            <i class="fa fa-vk"></i>
+                                        </a>
+                                    </div>
+                                    <!-- Social End -->
+                                </div>
+
+                            </div>
+                        </div>
+                    </div>
+                </div>
+
+                <span class="top-bar-arrow">
+                    <span>
+                        <i class="fa fa-angle-down"></i>
+                    </span>
+                </span>
+
+                <!-- Top Bar End -->
+
+                <!-- Branding & Navigation -->
+                <div class="wf-wrap clearfix">
+
+                    <div id="branding" class="wf-td">
+                        <a href="index.php">
+                            <img src="img/logo-regular.png" alt="Logo">
+                        </a>
+                    </div>
+
+
+                    <nav id="navigation" class="wf-td">
+                        <!-- Mobile nav -->
+                        <div id="dl-menu">
+                            <a href="#" id="mobile-menu">
+                                <div class="mobile-button">
+                                    <span class="menu-open" style="display: inline;">Menu</span>
+                                    <span class="menu-back">Back</span>
+                                </div>
+                            </a>
+                            <div class="dl-container">
+                                <ul class="dl-menu">
+                                    <li class="act">
+                                        <a href="#" onclick="return false;">
+                                            <span class="dl-text">Home</span>
+                                            <span class="dl-icon">
+                                                <i class="fa fa-arrow-right"></i>
+                                            </span>
+                                        </a>
+                                    </li>
+                                    <li>
+                                        <a href="#" onclick="return false;">
+                                            <span class="dl-text">Pages</span>
+                                            <span class="dl-icon">
+                                                <i class="fa fa-arrow-right"></i>
+                                            </span>
+                                        </a>
+                                    </li>
+                                    <li>
+                                        <a href="#" onclick="return false;">
+                                            <span class="dl-text">Blog</span>
+                                            <span class="dl-icon">
+                                                <i class="fa fa-arrow-right"></i>
+                                            </span>
+                                        </a>
+                                    </li>
+                                    <li>
+                                        <a href="#" onclick="return false;">
+                                            <span class="dl-text">Portfolio</span>
+                                            <span class="dl-icon">
+                                                <i class="fa fa-arrow-right"></i>
+                                            </span>
+                                        </a>
+                                    </li>
+                                    <li>
+                                        <a href="#" onclick="return false;">
+                                            <span class="dl-text">Gallery</span>
+                                            <span class="dl-icon">
+                                                <i class="fa fa-arrow-right"></i>
+                                            </span>
+                                        </a>
+                                    </li>
+                                </ul>
+                            </div>
+                        </div>
+
+                        <!-- Main Nav -->
+                        <ul id="main-nav" class="clearfix">
+                            <li class="act no-link">
+                                <a href="#" onclick="return false;" class="topItem">
+                                    <span>
+                                        Home
+                                    </span>
+                                </a>
+                                <div class="sub-nav">
+                                    <ul>
+                                        <li>
+                                            <a href="index.php" onclick="return false;">
+                                                <span>1. Business</span>
+                                            </a>
+                                        </li>
+                                        <li>
+                                            <a href="index.php" onclick="return false;">
+                                                <span>2. Presentation</span>
+                                            </a>
+                                        </li>
+                                        <li>
+                                            <a href="index.php" onclick="return false;">
+                                                <span>3. Creative</span>
+                                            </a>
+                                        </li>
+                                        <li>
+                                            <a href="index.php" onclick="return false;">
+                                                <span>4. Photography</span>
+                                            </a>
+                                        </li>
+                                    </ul>
+                                </div>
+                            </li>
+                            <li class="no-link">
+                                <a href="#" onclick="return false;" class="topItem">
+                                    <span>
+                                        Pages
+                                        <i class="underline"></i>
+                                    </span>
+                                </a>
+                                <div class="sub-nav">
+                                    <ul>
+                                        <li>
+                                            <a href="#" onclick="return false;">
+                                                <span>1. Business</span>
+                                            </a>
+                                        </li>
+                                        <li>
+                                            <a href="#" onclick="return false;">
+                                                <span>2. Presentation</span>
+                                            </a>
+                                        </li>
+                                        <li>
+                                            <a href="#" onclick="return false;">
+                                                <span>3. Creative</span>
+                                            </a>
+                                        </li>
+                                        <li>
+                                            <a href="#" onclick="return false;">
+                                                <span>4. Photography</span>
+                                            </a>
+                                        </li>
+                                    </ul>
+                                </div>
+                            </li>
+                            <li class="no-link">
+                                <a href="#" onclick="return false;" class="topItem">
+                                    <span>
+                                        Blog
+                                        <i class="underline"></i>
+                                    </span>
+                                </a>
+                                <div class="sub-nav">
+                                    <ul>
+                                        <li>
+                                            <a href="#" onclick="return false;">
+                                                <span>1. Business</span>
+                                            </a>
+                                        </li>
+                                        <li>
+                                            <a href="#" onclick="return false;">
+                                                <span>2. Presentation</span>
+                                            </a>
+                                        </li>
+                                        <li>
+                                            <a href="#" onclick="return false;">
+                                                <span>3. Creative</span>
+                                            </a>
+                                        </li>
+                                        <li>
+                                            <a href="#" onclick="return false;">
+                                                <span>4. Photography</span>
+                                            </a>
+                                        </li>
+                                    </ul>
+                                </div>
+                            </li>
+                            <li class="no-link">
+                                <a href="#" onclick="return false;" class="topItem">
+                                    <span>
+                                        Portfolio
+                                        <i class="underline"></i>
+                                    </span>
+                                </a>
+                                <div class="sub-nav">
+                                    <ul>
+                                        <li>
+                                            <a href="#" onclick="return false;">
+                                                <span>1. Business</span>
+                                            </a>
+                                        </li>
+                                        <li>
+                                            <a href="#" onclick="return false;">
+                                                <span>2. Presentation</span>
+                                            </a>
+                                        </li>
+                                        <li>
+                                            <a href="#" onclick="return false;">
+                                                <span>3. Creative</span>
+                                            </a>
+                                        </li>
+                                        <li>
+                                            <a href="#" onclick="return false;">
+                                                <span>4. Photography</span>
+                                            </a>
+                                        </li>
+                                    </ul>
+                                </div>
+                            </li>
+                            <li class="no-link">
+                                <a href="#" onclick="return false;" class="topItem">
+                                    <span>
+                                        Gallery
+                                        <i class="underline"></i>
+                                    </span>
+                                </a>
+                                <div class="sub-nav">
+                                    <ul>
+                                        <li>
+                                            <a href="#" onclick="return false;">
+                                                <span>1. Business</span>
+                                            </a>
+                                        </li>
+                                        <li>
+                                            <a href="#" onclick="return false;">
+                                                <span>2. Presentation</span>
+                                            </a>
+                                        </li>
+                                        <li>
+                                            <a href="#" onclick="return false;">
+                                                <span>3. Creative</span>
+                                            </a>
+                                        </li>
+                                        <li>
+                                            <a href="#" onclick="return false;">
+                                                <span>4. Photography</span>
+                                            </a>
+                                        </li>
+                                    </ul>
+                                </div>
+                            </li>
+                        </ul>
+                    </nav>
+
+                </div>
+                <!-- Branding & Navigation End -->
+
+            </header>
+
+            <!-- Header End -->
+
+            <!-- Page title -->
+
+            <div class="page-title">
+                <div class="wf-wrap">
+
+                    <div class="wf-table">
+                        <div class="wf-td hgroup">
+                            <h1>Вход</h1>
+                        </div>
+                        <div class="wf-td">
+                            <ul class="breadcrumbs text-normal">
+                                <li>
+                                    <a href="index.php">Главная</a>
+                                </li>
+                                <li>
+                                    <a href="login.php">Вход</a>
+                                </li>
+                            </ul>
+                        </div>
+                    </div>
+
+                </div>
+            </div>
+
+            <!-- Page title -->
+
+            <!-- Main Content -->
+
+            <div id="main">
+                <div class="container wf-wrap">
+                    <div class="row">
+
+                        <div class="form-container">
+                            <form name="login-form" action="#" method="POST">
+                                <div class="form-fields">
+                                    <span class="form-name login-form-input">
+                                        <input type="text" class="validate" placeholder="E-mail *" name="email" value="">
+                                    </span>
+                                    <span class="form-name login-form-input">
+                                        <input type="text" class="validate" placeholder="Пароль *" name="password" value="">
+                                    </span>
+                                </div>
+
+                                <p class="remember-me">
+                                    <label>
+                                        <input name="rememberme" type="checkbox" id="rememberme" value="">
+                                        Запомнить меня
+                                    </label>
+                                </p>
+
+                                <p>
+                                    <!--<a href="#" id="doRegistr" class="form-submit">-->                                    
+                                        <span>
+                                            <input type="submit" class="form-submit" name="login" value="Войти">
+                                        </span>
+                                    <!--</a>-->
+                                </p>
+                            </form>
+                            <?php
+                            require_once 'application/classes/user.php';
+                            $user = new User();
+                            if (isset($_POST['register'])) {
+                                $user->login();
+                            }
+                            $user->login();
+                            ?>﻿
+                        </div>
+
+                    </div>
+
+                    <hr>
+
+                    <div class="row">
+                        <!-- Feedback Section -->
+
+                        <div id="feedback-container">
+
+                            <header>
+                                <h2>Наши клиенты и партнеры</h2>
+                                <h4>Нас уже более 10 000!</h4>
+                            </header>
+
+                            <div class="feeback-content clearfix">
+
+                                <div id="feedback-carousel-comments" class="col-sm-6">
+
+                                    <!-- Customer comments block -->
+                                    <section id="testimonial-slider" class="carousel slide">
+
+                                        <ol class="carousel-indicators">
+                                            <li data-target="#testimonial-slider" data-slide-to="0" class="active"></li>
+                                            <li data-target="#testimonial-slider" data-slide-to="1"></li>
+                                            <li data-target="#testimonial-slider" data-slide-to="2"></li>
+                                        </ol>
+
+                                        <div class="carousel-inner" role="listbox">
+                                            <div class="item active">
+                                                <p>
+                                                    Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris ipsum erat, finibus sit amet fringilla id, accumsan sed nisl. 
+                                                    Nulla odio eros, blandit ac metus faucibus, sollicitudin posuere sapien. Maecenas ut convallis arcu. Phasellus at tellus 
+                                                    sed odio vestibulum sodales in at lacus. Sed commodo metus et sapien pretium, ac sollicitudin enim dignissim. Donec tempus 
+                                                    diam et porta aliquam. Ut efficitur sollicitudin diam a accumsan.
+                                                </p>
+                                                <span class="img">
+                                                    <img src="img/feedback-person-1.jpg" alt="person 1">
+                                                </span>
+                                                <span class="who">
+                                                    <span class="text-primary">Diana Richards</span>
+                                                    <br>
+                                                    <span class="text-secondary">manager</span>
+                                                </span>
+                                            </div>
+                                            <div class="item">
+                                                <p>
+                                                    Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris ipsum erat, finibus sit amet fringilla id, accumsan sed nisl. 
+                                                    Nulla odio eros, blandit ac metus faucibus, sollicitudin posuere sapien. Maecenas ut convallis arcu. Phasellus at tellus 
+                                                    sed odio vestibulum sodales in at lacus. Sed commodo metus et sapien pretium, ac sollicitudin enim dignissim. Donec tempus 
+                                                    diam et porta aliquam. Ut efficitur sollicitudin diam a accumsan.
+                                                </p>
+                                                <span class="img">
+                                                    <img src="img/feedback-person-2.jpg" alt="person 2">
+                                                </span>
+                                                <span class="who">
+                                                    <span class="text-primary">Anna White</span>
+                                                    <br>
+                                                    <span class="text-secondary">loyal client</span>
+                                                </span>
+                                            </div>
+                                            <div class="item">
+                                                <p>
+                                                    Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris ipsum erat, finibus sit amet fringilla id, accumsan sed nisl. 
+                                                    Nulla odio eros, blandit ac metus faucibus, sollicitudin posuere sapien. Maecenas ut convallis arcu. Phasellus at tellus 
+                                                    sed odio vestibulum sodales in at lacus. Sed commodo metus et sapien pretium, ac sollicitudin enim dignissim. Donec tempus 
+                                                    diam et porta aliquam. Ut efficitur sollicitudin diam a accumsan.
+                                                </p>
+                                                <span class="img">
+                                                    <img src="img/feedback-person-3.jpg" alt="person 3">
+                                                </span>
+                                                <span class="who">
+                                                    <span class="text-primary">Jacob Firebird</span>
+                                                    <br>
+                                                    <span class="text-secondary">happy client</span>
+                                                </span>
+                                            </div>
+                                        </div>
+
+                                    </section>
+                                    <!-- Customer comments block End -->
+                                </div>
+
+                                <div id="feedback-partners" class="col-sm-6">
+                                    <!-- Partners icons block -->
+                                    <section id="partners-icons" class="clearfix">
+
+                                        <div class="col-sm-4 partner-icon">
+                                            <a href="#" target="_blank" title="Partner 1">
+                                                <img src="img/l-authentic.png">
+                                            </a>
+                                        </div>
+
+                                        <div class="col-sm-4 partner-icon">
+                                            <a href="#" target="_blank" title="Partner 2">
+                                                <img src="img/l-hamb.png">
+                                            </a>
+                                        </div>
+
+                                        <div class="col-sm-4 partner-icon">
+                                            <a href="#" target="_blank" title="Partner 3">
+                                                <img src="img/l-ransom.png">
+                                            </a>
+                                        </div>
+
+                                        <div class="col-sm-4 partner-icon">
+                                            <a href="#" target="_blank" title="Partner 3">
+                                                <img src="img/l-steak.png">
+                                            </a>
+                                        </div>
+
+                                        <div class="col-sm-4 partner-icon">
+                                            <a href="#" target="_blank" title="Partner 3">
+                                                <img src="img/l-thesaviour.png">
+                                            </a>
+                                        </div>
+
+                                        <div class="col-sm-4 partner-icon">
+                                            <a href="#" target="_blank" title="Partner 3">
+                                                <img src="img/l-vintage.png">
+                                            </a>
+                                        </div>
+
+                                    </section>
+                                    <!-- Partners icons block End -->
+                                </div>
+
+                            </div>
+
+                        </div>
+
+                        <!-- Feedback Section End -->
+                    </div>
+
+
+                </div>
+            </div>
+
+            <!-- Main Content End -->
+
+            <footer id="footer">
+
+                <div class="wf-wrap">
+                    <div class="wf-container-footer">
+                        <div class="wf-container clearfix">
+
+                            <section class="widget col-sm-4" id="footer-contact-us-widget">
+                                <div class="widget-title">НАШИ КОНТАКТЫ</div>
+                                <ul class="contact-info">
+                                    <li>
+                                        <span class="color-primary">Адрес:</span>
+                                        <br>
+                                        8500 Beverly Boulevard
+                                        <br>
+                                        Los Angeles, CA 90048
+                                    </li>
+                                    <li>
+                                        <span class="color-primary">Телефоны:</span>
+                                        <br>
+                                        +12 345 67 00 89
+                                        <br>
+                                        +12 987 00 65 43
+                                    </li>
+                                    <li>
+                                        <span class="color-primary">E-mail:</span>
+                                        <br>
+                                        isitinme88@gmail.com
+                                    </li>
+                                </ul>
+                                <div class="soc-ico clearfix">
+                                    <a href="#" title="Instagram" target="_blank">
+                                        <i class="fa fa-instagram"></i>
+                                    </a>
+                                    <a href="#" title="Facebook" target="_blank">
+                                        <i class="fa fa-facebook-square"></i>
+                                    </a>
+                                    <a href="#" title="Twitter" target="_blank">
+                                        <i class="fa fa-twitter"></i>
+                                    </a>
+                                    <a href="#" title="Vk" target="_blank">
+                                        <i class="fa fa-vk"></i>
+                                    </a>
+                                </div>
+                            </section>
+
+                            <section class="widget col-sm-4" id="footer-our-services-widget">
+                                <div class="widget-title">НАШИ ПРЕДЛОЖЕНИЯ</div>
+
+                                <ul class="foot-accordion clearfix">
+                                    <li>
+                                        <a href="#" class="text-primary">
+                                            <span class="accor-toggle-icon">
+                                                <i class="fa fa-angle-down"></i>
+                                            </span>
+                                            <span class="color-primary">Web Design</span>
+                                        </a>
+                                        <div class="accor-content">
+                                            Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris ipsum erat, finibus sit amet fringilla id, accumsan sed nisl. 
+                                        </div>
+                                    </li>
+                                    <li>
+                                        <a href="#" class="text-primary">
+                                            <span class="accor-toggle-icon">
+                                                <i class="fa fa-angle-down"></i>
+                                            </span>
+                                            <span class="color-primary">Photography</span>
+                                        </a>
+                                        <div class="accor-content">
+                                            Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris ipsum erat, finibus sit amet fringilla id, accumsan sed nisl. 
+                                        </div>
+                                    </li>
+                                    <li>
+                                        <a href="#" class="text-primary">
+                                            <span class="accor-toggle-icon">
+                                                <i class="fa fa-angle-down"></i>
+                                            </span>
+                                            <span class="color-primary">Programming</span>
+                                        </a>
+                                        <div class="accor-content">
+                                            Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris ipsum erat, finibus sit amet fringilla id, accumsan sed nisl. 
+                                        </div>
+                                    </li>
+                                    <li>
+                                        <a href="#" class="text-primary">
+                                            <span class="accor-toggle-icon">
+                                                <i class="fa fa-angle-down"></i>
+                                            </span>
+                                            <span class="color-primary">Marketing & PR</span>
+                                        </a>
+                                        <div class="accor-content">
+                                            Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris ipsum erat, finibus sit amet fringilla id, accumsan sed nisl. 
+                                        </div>
+                                    </li>
+                                </ul>
+
+                            </section>
+
+                            <section class="widget col-sm-4" id="footer-featured-posts-widget">
+                                <div class="widget-title">ПОПУЛЯРНЫЕ СТАТЬИ</div>
+
+                                <ul class="recent-posts">
+                                    <li>
+                                        <article>
+                                            <div class="wf-td">
+                                                <a href="#" class="alignLeft">
+                                                    <img src="img/article_img_1.jpg" alt="Image" width="40" height="40">
+                                                </a>
+                                            </div>
+                                            <div class="post-content">
+                                                <a href="#">Business cards design trends</a>
+                                                <br>
+                                                <time class="text-secondary">August 5, 2013</time>
+                                            </div>
+                                        </article>
+                                    </li>
+                                    <li>
+                                        <article>
+                                            <div class="wf-td">
+                                                <a href="#" class="alignLeft">
+                                                    <img src="img/article_img_1.jpg" alt="Image" width="40" height="40">
+                                                </a>
+                                            </div>
+                                            <div class="post-content">
+                                                <a href="#">Business cards design trends</a>
+                                                <br>
+                                                <time class="text-secondary">August 5, 2013</time>
+                                            </div>
+                                        </article>
+                                    </li>
+                                    <li>
+                                        <article>
+                                            <div class="wf-td">
+                                                <a href="#" class="alignLeft">
+                                                    <img src="img/article_img_1.jpg" alt="Image" width="40" height="40">
+                                                </a>
+                                            </div>
+                                            <div class="post-content">
+                                                <a href="#">Business cards design trends</a>
+                                                <br>
+                                                <time class="text-secondary">August 5, 2013</time>
+                                            </div>
+                                        </article>
+                                    </li>
+                                    <li>
+                                        <article>
+                                            <div class="wf-td">
+                                                <a href="#" class="alignLeft">
+                                                    <img src="img/article_img_1.jpg" alt="Image" width="40" height="40">
+                                                </a>
+                                            </div>
+                                            <div class="post-content">
+                                                <a href="#">Business cards design trends</a>
+                                                <br>
+                                                <time class="text-secondary">August 5, 2013</time>
+                                            </div>
+                                        </article>
+                                    </li>
+                                </ul>
+                            </section>
+
+                        </div>
+                    </div>
+                </div>
+
+            </footer>
+
+            <a href="#" class="scroll-top">
+                <i class="fa fa-arrow-up"></i>
+            </a>
+
+        </section>
+
+
+        <script src="js/jquery-1.11.0.min.js"></script>
+        <script src="js/jquery-ui.js"></script>
+        <script src="js/bootstrap.js"></script>
+        <script src="js/perfect-scrollbar.jquery.js"></script>
+        <script src="js/jquery.fancybox.pack.js"></script>
+        <script src="js/fixedMenu.js"></script>
+        <script src="js/main.js"></script>
+    </body>
+</html>
